requirements.txt
asgiref==3.8.1
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
pgsql==2.2
psycopg==3.2.9
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
